+++ next.config.js
@@ -1,0 +1,9 @@
+const nextConfig = {
+  reactStrictMode: true,
+  swcMinify: true,
+  experimental: {
+    scrollRestoration: true,
+  },
+}
+
+module.exports = nextConfig
